+++ crates/wdl-analysis/src/analyzer.rs
@@ -22,6 +22,7 @@ use line_index::WideEncoding;
 use line_index::WideLineCol;
 use lsp_types::CompletionResponse;
 use lsp_types::GotoDefinitionResponse;
+use lsp_types::Hover;
 use lsp_types::Location;
 use path_clean::PathClean;
 use tokio::runtime::Handle;
@@ -41,6 +42,7 @@ use crate::queue::CompletionRequest;
 use crate::queue::FindAllReferencesRequest;
 use crate::queue::FormatRequest;
 use crate::queue::GotoDefinitionRequest;
+use crate::queue::HoverRequest;
 use crate::queue::NotifyChangeRequest;
 use crate::queue::NotifyIncrementalChangeRequest;
 use crate::queue::RemoveRequest;
@@ -674,6 +676,32 @@ where
                 "failed to send completion request to analysis queue because the channel has \
                  closed"
             )
+        })
+    }
+
+    /// Performs a `hover` for a symbol at a given position in a document.
+    pub async fn hover(
+        &self,
+        document: Url,
+        position: SourcePosition,
+        encoding: SourcePositionEncoding,
+    ) -> Result<Option<Hover>> {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::Hover(HoverRequest {
+                document,
+                position,
+                encoding,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!(
+                    "failed to send hover request to analysis queue because the channel has closed"
+                )
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!("failed to send hover request to analysis queue because the channel has closed")
         })
     }
 }