+++ crates/wdl-ast/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Fixed
+
+* Workflow markdown styling ([#540](https://github.com/stjude-rust-labs/wdl/pull/540)).
+
 ## 0.15.1 - 07-31-2025
 
 #### Added